LayoutTests/platform/chromium/fast/js/instance-of-immediates-expected.txt

This test makes sure that instance of behaves correctly when the value, constructor, or its prototype are immediates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (1 instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got 1.
PASS ({} instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got #<Object>.
PASS (obj instanceof 1) threw exception TypeError: Expecting a function in instanceof check, but got #<Constructor>.
PASS (1 instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got 1.
PASS ({} instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got #<Object>.
PASS (obj instanceof {}) threw exception TypeError: Expecting a function in instanceof check, but got #<Constructor>.
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) is false
PASS (obj instanceof Constructor) is true
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) threw exception TypeError: Function has non-object prototype '1' in instanceof check.
PASS (obj instanceof Constructor) threw exception TypeError: Function has non-object prototype '1' in instanceof check.
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) is false
PASS (obj instanceof Constructor) is false
PASS (1 instanceof Constructor) is false
PASS ({} instanceof Constructor) threw exception TypeError: Function has non-object prototype 'null' in instanceof check.
PASS (obj instanceof Constructor) threw exception TypeError: Function has non-object prototype 'null' in instanceof check.
PASS successfullyParsed is true

TEST COMPLETE
